+++ RegularExpression/OGRegularExpressionEnumeratorPrivate.m
@@ -49,7 +49,7 @@
         /* DEBUG 
         {
             NSLog(@"TargetString: '%@'", _targetString);
-            int     i, count = _lengthOfTargetString;
+            NSInteger i, count = _lengthOfTargetString;
             unichar *utf16Chars = _UTF16TargetString;
             for (i = 0; i < count; i++) {
                 NSLog(@"UTF16: %04x", *(utf16Chars + i));
